Add order control strip
The new widget is a row of buttons that perform actions on an order.
So far onpress for only paid and receipt buttons have been implemented.

Since these buttons modify a `CustomerOrder' object's properties,
we have made those properties non-final.

Lastly, `OrderItemCard` widget's is-done button was extracted
to reuse it as order action buttons.