Improve testability of classes that use Firebase
FirebaseService has been converted from a collection of static methods
to a class with regular instance methods. This makes it easy to pass
specialized instances of FirebaseService (eg. mocks)  whenever required.

Consequently, we've refactored all widgets where FirebaseService is
being used to accept an instance of FirebaseService via constructor.

Additionally, we've created a MockFirebaseService class for use in unit
or widget tests.